+++ examples/canvasdemo.jl
@@ -1,11 +1,8 @@
 using Colors, TclTk
 
-# Make sure Tk package is loaded and event loop is running.
-interp = tk_start()
-
 # Top-level widget with title.
 top = Toplevel(background="#282c34")
-wm.title(top, "Canvas demo")
+tk_wm.title(top, "Canvas demo")
 
 # Create a canvas widget. The "closeenough" settings (in pixels) is to facilitate the
 # selection of a marker to delete.
@@ -20,7 +17,7 @@ Click on an empty location to add a marker.
 Click on a marker to delete it.""", aspect=600)
 
 # Create a Tcl variable to track the number of markers.
-counter = TclTk.Variable{Int}(interp, "::NUMBER_OF_MARKERS")
+counter = TclVariable{Int}(interp, "::NUMBER_OF_MARKERS")
 counter[] = 0
 
 # Create a label to display the number of markers.
@@ -53,9 +50,9 @@ function add_marker(canvas::Canvas, xm, ym;
 end
 
 # Callback function to be called with: %W %x %y.
-function on_click(interp::TclInterp, args::TclObj)
+function on_click(args::TclObj)
     # Extract arguments (1st is name of procedure, unused here).
-    canvas, xm, ym = interp.fetch(Tuple{Canvas,Float64,Float64}, args, 2:4)
+    canvas, xm, ym = fetch(Tuple{Canvas,Float64,Float64}, args, 2:4)
 
     # Convert coordinates to canvas coordinates.
     x = canvas.canvasx(Float64, xm)
@@ -77,7 +74,7 @@ function on_click(interp::TclInterp, args::TclObj)
 end
 
 # Create the counterpart of the callback in the Tcl interpreter.
-on_click_callback = TclTk.Callback(on_click, interp, "on_click")
+on_click_callback = TclCallback(on_click, interp, "on_click")
 
 # Bind event to callback.
 bind(canvas, "<ButtonPress-1>", "on_click %W %x %y")